+++ NEWS.md
@@ -13,6 +13,7 @@ Enhancements:
   * Never show unknown events during movement table display.
   * Show how many invalid events were omitted in the message that opens movement events.
   * Expand unknown receiver options to allow discarding unknown detections and repeating the same action for all occurrences.
+  * Improve error messages when something fails during detections loading.
 
 ## actel 1.1.0